% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defries-fulker.R
\name{DeFriesFulkerMethod1}
\alias{DeFriesFulkerMethod1}
\alias{DeFriesFulkerMethod3}
\title{Estimates the heritability of additive traits using DeFries-Fulker Method 1.}
\usage{
DeFriesFulkerMethod1(dataSet, oName_S1, oName_S2, rName = "R")

DeFriesFulkerMethod3(dataSet, oName_S1, oName_S2, rName = "R")
}
\description{
Estimates the heritability of additive traits using DeFries-Fulker Method 1.

Estimates the heritability of additive traits using DeFries-Fulker Method 3.
}
