\name{CreateSpatialNeighbours79Gen2}
\alias{CreateSpatialNeighbours79Gen2}
\title{Automatically creates dataset of NLSY79 Gen2 subjects (i.e., the children of the mothers in the initial NLSY79 sample).}
\usage{
  CreateSpatialNeighbours79Gen2()
}
\description{
  \code{CreateSpatialNeighbours79Gen2} automatically
  creates dataset of NLSY79 Gen2 subjects (i.e., the
  children of the mothers in the initial NLSY79 sample).

  For more information, please see
  \code{\link{CreateSpatialNeighbours}}
}

