% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ace-estimate.R
\docType{methods}
\name{GetDetails,AceEstimate-method}
\alias{GetDetails,AceEstimate-method}
\title{A generic function for extracting the \code{Details} slot of an object.}
\usage{
\S4method{GetDetails}{AceEstimate}(object)
}
\arguments{
\item{object}{ACE object}
}
\description{
A generic function for extracting the \code{Details} slot of an AceEstimation object.
}
\author{
Will Beasley
}
\keyword{methods}
