% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ace-estimate.R
\name{GetDetails-methods}
\alias{GetDetails-methods}
\alias{GetDetails}
\alias{AceEstimate-method}
\title{GetDetails-methods}
\usage{
GetDetails(object)
}
\arguments{
\item{object}{ACE object}
}
\description{
GetDetails-methods
}
