\docType{methods}
\name{GetDetails-methods}
\alias{AceEstimate-method}
\alias{GetDetails}
\alias{GetDetails-methods}
\title{A generic function for extracting the \code{Details} slot of an object.}
\description{
  A generic function for extracting the \code{Details} slot
  of an \code{AceEstimation} object.
}
\section{Methods
 \itemize{
    \item{list("signature(object = \"AceEstimate\")")}{Extracts the \code{Details} slot of an \code{AceEstimation} object.}
 }}{

}
\author{
  Will Beasley
}
\keyword{methods}

