% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{GetDetails-methods}
\alias{AceEstimate-method}
\alias{GetDetails}
\alias{GetDetails-methods}
\title{A generic function for extracting the \code{Details} slot of an object.}
\description{
A generic function for extracting the \code{Details} slot of an \code{AceEstimation} object.
}
\author{
Will Beasley
}
\keyword{methods}

