\name{NlsyLinks-package}
\alias{NlsyLinks-package}
\alias{NlsyLinks}
\docType{package}
\title{
Utilities and kinship information for Behavioral Genetics research using the NLSY
}
\description{
Utilities and kinship information for Behavioral Genetics research using the NLSY
}
\details{
\tabular{ll}{
Package: \tab NlsyLinks\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2012-01-15\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
An overview of how to use the package, including the most important functions
}
\author{
Will Beasley, Joe Rodgers, Kelly Williams, and David Bard

Maintainer: Will  Beasley <wibeasley@hotmail.com>
}
\references{
Rodgers, Joseph Lee, & Kohler, Hans-Peter (2005).  Reformulating and simplifying the DF analysis model.  \href{http://www.springerlink.com/content/n3x1v1q282583366/}{Behavior Genetics, 35 (2), 211-217}.
}

\keyword{ package }
\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
f(x=1, y=5)
}
